paste now places text at the cursor position; treeview got value and selected attributes, the first one selects a node by its label and the second one - by the data; #173 fix; playing around the calcvoid optimisation...

+++ xlib/asymmunit.cpp
@@ -635,7 +635,7 @@ double TAsymmUnit::CalcCellVolume()  const  {
           FAxes[1].GetV()*
           FAxes[2].GetV()*sqrt( (1-cosa*cosa-cosb*cosb-cosg*cosg) + 2*(cosa*cosb*cosg));
 }
-double TAsymmUnit::EstimateZ(size_t atomCount) const  {
+double TAsymmUnit::EstimateZ(double atomCount) const  {
   double auv = (double)(CalcCellVolume()/(TUnitCell::GetMatrixMultiplier(GetLatt())*(MatrixCount()+1)));
   int zp = olx_round(auv/(18.6*atomCount));
   return (double)olx_max((TUnitCell::GetMatrixMultiplier(GetLatt())*(MatrixCount()+1) * zp), 1);
